Created utils file and began separating gui code from sim code
