+++ start.py
@@ -3,16 +3,16 @@ from tkinter import messagebox
 from gui import UVSimGUI
 
 class UVSim:
-    def __init__(self, gui):
+    def __init__(self):
         self.memory = [0] * 100
         self.accumulator = 0
         self.pc = 0
         self.running = True
         self.waiting_for_input = False
-        self.gui = gui
+        self.gui = None
 
-    #def set_gui(self, gui):
-        #self.gui = gui
+    def set_gui(self, gui):
+        self.gui = gui
 
     def load_program(self, program):
         self.memory = [0] * 100
@@ -89,34 +89,10 @@ class Read(Operation):
     def execute(self):
         if self.sim.gui:
             self.sim.waiting_for_input = True
-            input_dialog = tk.Toplevel(self.sim.gui.root)
-            input_dialog.title("Input")
-            tk.Label(input_dialog, text=f"Enter an integer for memory location {self.operand}:").pack()
-            input_var = tk.IntVar()
-
-            def on_submit(event=None):
-                try:
-                    value = int(entry.get())
-                    input_var.set(value)
-                    input_dialog.destroy()
-                except ValueError:
-                    messagebox.showerror("Invalid input", "Please enter a valid integer.")
-
-            entry = tk.Entry(input_dialog)
-            entry.pack()
-            entry.bind("<Return>", on_submit)
-            tk.Button(input_dialog, text="Submit", command=on_submit).pack()
-            input_dialog.transient(self.sim.gui.root)
-            input_dialog.grab_set()
-            input_dialog.geometry(f"+{self.sim.gui.root.winfo_rootx() + self.sim.gui.root.winfo_width() // 2 - input_dialog.winfo_reqwidth() // 2}+{self.sim.gui.root.winfo_rooty() + self.sim.gui.root.winfo_height() // 2 - input_dialog.winfo_reqheight() // 2}")
-            self.sim.gui.root.wait_window(input_dialog)
-
-            value = input_var.get()
-            self.sim.memory[self.operand] = value
-            self.sim.waiting_for_input = False
-        else:
-            value = int(input(f"Enter an integer for memory location {self.operand}: "))
-            self.sim.memory[self.operand] = value
+            get_input = UVSimGUI.read_output(self.operand)
+            self.memory[get_input]
+        value = int(input(f"Enter an integer for memory location {self.operand}: "))
+        self.sim.memory[self.operand] = value
 
 class Write(Operation):
     def execute(self):
@@ -179,27 +155,14 @@ class Halt(Operation):
             self.sim.gui.display_message("Halting execution")
         self.sim.running = False
 
-def load_program_from_file(filename):
-    program = []
-    with open(filename, 'r') as file:
-        for line in file:
-            line = line.strip()
-            if line:
-                if UVSim.is_valid_instruction(line):
-                    program.append(int(line))
-                else:
-                    print(f"Invalid instruction '{line}' ignored.")
-    return program
-
 def main():
  root = tk.Tk()
  interface = UVSimGUI(root)
+ get_file = interface.load_file()
  app = UVSim()
- load_program_from_file()
- app.load_program()
- app.fetch()
- root.mainloop()
+ app.load_program(get_file)
+ app.run()
+ 
 
 if __name__ == "__main__":
-    import gui
-    gui.main()
+    main()
