+++ utils.py
@@ -1,0 +1,6 @@
+def load_program_from_file(filename):
+    program = []
+    with open(filename, 'r') as file:
+        for line in file:
+            line = line.strip()
+    return program
